Fix dtype mismatch in scatter_ operations
Mixed precision training에서 dtype 불일치로 인한 에러 수정:
- scatter_ 연산 시 source tensor의 dtype을 destination과 일치시킴
- forward()와 get_neuron_stats() 모두 수정
- input_repr 생성 시 명시적으로 dtype 지정

Error fixed:
RuntimeError: scatter(): Expected self.dtype to be equal to src.dtype

+++ models/three_stage_ffn.py
@@ -137,7 +137,8 @@ class ThreeStageFFN(nn.Module):
             device=x.device,
             dtype=x.dtype
         )
-        input_repr.scatter_(2, top_input_idx, top_input_acts)
+        # Ensure dtype compatibility for scatter operation
+        input_repr.scatter_(2, top_input_idx, top_input_acts.to(input_repr.dtype))
         # [B, S, n_input]
 
         # 처리 뉴런 활성화
@@ -211,8 +212,8 @@ class ThreeStageFFN(nn.Module):
             top_input_acts, top_input_idx = input_acts.topk(k_input, dim=-1)
 
             # Stage 2
-            input_repr = torch.zeros(B, S, self.n_input, device=x.device)
-            input_repr.scatter_(2, top_input_idx, top_input_acts)
+            input_repr = torch.zeros(B, S, self.n_input, device=x.device, dtype=x.dtype)
+            input_repr.scatter_(2, top_input_idx, top_input_acts.to(input_repr.dtype))
             process_acts = F.gelu(input_repr @ self.process_input_weights.T)
             top_process_acts, top_process_idx = process_acts.topk(k_process, dim=-1)
